+++ src/Futhark/CLI/Test.hs
@@ -144,9 +144,9 @@ instance Eq TestCase where
 instance Ord TestCase where
   x `compare` y = testCaseProgram x `compare` testCaseProgram y
 
-data RunResult
+data RunResult m
   = ErrorResult T.Text
-  | SuccessResult [Value]
+  | SuccessResult (m [Value]) (m ())
 
 progNotFound :: T.Text -> T.Text
 progNotFound s = s <> ": command not found"
@@ -247,8 +247,8 @@ runInterpretedEntry (FutharkExe futhark) program (InputOutputs entry run_cases) 
                 throwError $ progNotFound $ T.pack futhark
               _ ->
                 liftExcept $
-                  compareResult entry index program expectedResult'
-                    =<< runResult program code output err
+                  compareResult entry index program expectedResult' $
+                    runResult program code output err
    in accErrors_ $ map runInterpretedCase run_cases
 
 runTestCase :: TestCase -> TestM ()
@@ -366,13 +366,13 @@ runCompiledEntry futhark server program (InputOutputs entry run_cases) = do
         call_r <- liftIO $ cmdCall server entry outs ins
         liftCommand $ cmdFree server ins
 
-        res <- case call_r of
-          Left (CmdFailure _ err) ->
-            pure $ ErrorResult $ T.unlines err
-          Right _ ->
-            SuccessResult
-              <$> readResults server outs
-              <* liftCommand (cmdFree server outs)
+        let res = case call_r of
+              Left (CmdFailure _ err) ->
+                ErrorResult $ T.unlines err
+              Right _ ->
+                SuccessResult
+                  (readResults server outs <* liftCommand (cmdFree server outs))
+                  (liftCommand (cmdFree server outs))
 
         compareResult entry index program expected res
 
@@ -393,16 +393,18 @@ runResult ::
   ExitCode ->
   SBS.ByteString ->
   SBS.ByteString ->
-  m RunResult
+  RunResult m
 runResult program ExitSuccess stdout_s _ =
-  case valuesFromByteString "stdout" $ LBS.fromStrict stdout_s of
-    Left e -> do
-      let actualf = program `addExtension` "actual"
-      liftIO $ SBS.writeFile actualf stdout_s
-      E.throwError $ T.pack e <> "\n(See " <> T.pack actualf <> ")"
-    Right vs -> pure $ SuccessResult vs
+  SuccessResult fetch (pure ())
+  where
+    fetch = case valuesFromByteString "stdout" $ LBS.fromStrict stdout_s of
+      Left e -> do
+        let actualf = program `addExtension` "actual"
+        liftIO $ SBS.writeFile actualf stdout_s
+        E.throwError $ T.pack e <> "\n(See " <> T.pack actualf <> ")"
+      Right vs -> pure vs
 runResult _ (ExitFailure _) _ stderr_s =
-  pure $ ErrorResult $ T.decodeUtf8 stderr_s
+  ErrorResult $ T.decodeUtf8 stderr_s
 
 compileTestProgram :: [String] -> FutharkExe -> String -> FilePath -> [WarningTest] -> TestM ()
 compileTestProgram extra_options futhark backend program warnings = do
@@ -417,11 +419,12 @@ compareResult ::
   Int ->
   FilePath ->
   ExpectedResult [Value] ->
-  RunResult ->
+  RunResult m ->
   m ()
-compareResult _ _ _ (Succeeds Nothing) SuccessResult {} =
-  pure ()
-compareResult entry index program (Succeeds (Just expected_vs)) (SuccessResult actual_vs) =
+compareResult _ _ _ (Succeeds Nothing) (SuccessResult _ free_vs) =
+  free_vs
+compareResult entry index program (Succeeds (Just expected_vs)) (SuccessResult get_actual_vs _) = do
+  actual_vs <- get_actual_vs
   checkResult
     (program <.> T.unpack entry <.> show index)
     expected_vs
@@ -430,7 +433,8 @@ compareResult _ _ _ (RunTimeFailure expectedError) (ErrorResult actualError) =
   checkError expectedError actualError
 compareResult _ _ _ (Succeeds _) (ErrorResult err) =
   E.throwError $ "Function failed with error:\n" <> err
-compareResult _ _ _ (RunTimeFailure f) (SuccessResult _) =
+compareResult _ _ _ (RunTimeFailure f) (SuccessResult _ free_vs) = do
+  free_vs
   E.throwError $ "Program succeeded, but expected failure:\n  " <> showText f
 
 ---